[custom] dependency for custom_engine
