+++ paddle/fluid/framework/new_executor/instruction/control_flow/if_instruction.h
@@ -16,6 +16,7 @@
 
 #include "paddle/fluid/framework/new_executor/instruction/instruction_base.h"
 #include "paddle/fluid/framework/new_executor/interpreter/execution_config.h"
+#include "paddle/fluid/framework/new_executor/new_executor_defs.h"
 
 namespace ir {
 class Operation;
