+++ paddle/fluid/framework/new_executor/instruction/custom_engine_instruction.h
@@ -15,8 +15,8 @@
 #pragma once
 #include "paddle/fluid/custom_engine/custom_engine_ext.h"
 #include "paddle/fluid/framework/new_executor/instruction/instruction_base.h"
-#include "paddle/fluid/framework/new_executor/pir_adaptor/pir_adaptor_util.h"
-#include "paddle/phi/core/platform/device_context.h"
+// #include "paddle/fluid/framework/new_executor/pir_adaptor/pir_adaptor_util.h"
+// #include "paddle/phi/core/platform/device_context.h"
 namespace pir {
 class Operation;
 }  // namespace pir
