+++ paddle/fluid/framework/new_executor/new_executor_defs.h
@@ -19,6 +19,7 @@
 #include <vector>
 
 #include "paddle/fluid/framework/infershape_utils.h"
+#include "paddle/fluid/framework/new_executor/new_executor_base_defs.h"
 #include "paddle/fluid/framework/operator.h"
 #include "paddle/fluid/framework/variable_helper.h"
 #include "paddle/fluid/pir/dialect/operator/interface/infermeta.h"
@@ -46,18 +47,6 @@ using OpKernelComputeFunc = std::function<void(const ExecutionContext&)>;
 using HookFunc = std::function<void(OperatorBase*, Scope*)>;
 using PirHookFunc =
     std::function<void(InstructionBase*, ValueExecutionInfo*, Scope*)>;
-
-using SchedulingPriority = int64_t;
-
-constexpr const char* kCoalesceTensor = "coalesce_tensor";
-
-// stream types
-constexpr const char* kCustomStream = "CustomStream";
-constexpr const char* kDefaultStream = "DefaultStream";
-constexpr const char* kD2HStream = "D2HStream";
-constexpr const char* kH2DStream = "H2DStream";
-
-constexpr int kEmptyVarIndex = 0;
 
 struct OpKernelFunc {
   OpKernelComputeFunc compute_func_;
@@ -149,22 +138,6 @@ class VariableScope {
 
   // var_name -> var_type
   std::vector<std::pair<std::string, int>> data_transfer_added_vars_;
-};
-
-struct EventInter {
-  explicit EventInter(size_t instr_id,
-                      std::shared_ptr<platform::DeviceEvent> event,
-                      platform::DeviceType waiter_type)
-      : instr_id_(instr_id), event_(event), waiter_type_(waiter_type) {}
-  size_t instr_id_;
-  std::shared_ptr<platform::DeviceEvent> event_;
-  platform::DeviceType waiter_type_;
-};
-
-enum class OpFuncType {
-  kCpuSync,  // CPU kernel, block host
-  kGpuSync,  // GPU or other device kernel without asynchronous operation
-  kGpuAsync  // GPU or other device kernel with asynchronous operation
 };
 
 struct OpFuncNode {
