+++ paddle/fluid/ir_adaptor/translator/utils.h
@@ -20,36 +20,10 @@
 #include "paddle/fluid/framework/data_type.h"
 #include "paddle/fluid/framework/program_desc.h"
 #include "paddle/fluid/ir_adaptor/translator/program_translator.h"
+#include "paddle/fluid/ir_adaptor/translator/utils_defs.h"
 #include "paddle/pir/include/core/ir_context.h"
 #include "paddle/pir/include/core/operation.h"
 #include "paddle/pir/include/core/program.h"
-
-namespace paddle {
-namespace dialect {
-struct PdOpSig {
-  std::string name;
-  std::vector<std::string> inputs;
-  std::vector<std::string> outputs;
-  PdOpSig() = default;
-  PdOpSig(const PdOpSig& input_info) = default;
-
-  PdOpSig(const std::string& name,
-          const std::vector<std::string>& inputs,
-          const std::vector<std::string>& outputs)
-      : name(name), inputs(inputs), outputs(outputs) {}
-};
-
-bool HaveOpToMultiKernelsMap(std::string op_name);
-
-const std::vector<PdOpSig>& LegacyOpToPdOpsMapping(std::string op_name);
-const std::vector<PdOpSig>& SparseOpToPdOpsMapping(std::string op_name);
-
-#ifdef PADDLE_WITH_DNNL
-bool IsOneDNNOnlyOp(std::string op_name);
-#endif
-
-}  // namespace dialect
-}  // namespace paddle
 
 namespace paddle {
 namespace translator {
