+++ paddle/fluid/pir/dialect/op_generator/op_gen.py
@@ -113,8 +113,9 @@ H_FILE_TEMPLATE = """// This file is generated by "paddle/fluid/pir/dialect/op_g
 #ifdef PADDLE_WITH_DNNL
 #include "paddle/fluid/pir/dialect/operator/trait/onednn.h"
 #endif
-#include "paddle/fluid/framework/infershape_utils.h"
-#include "paddle/fluid/ir_adaptor/translator/utils.h"
+// #include "paddle/fluid/framework/infershape_utils.h"
+// #include "paddle/fluid/ir_adaptor/translator/utils.h"
+#include "paddle/fluid/ir_adaptor/translator/utils_defs.h"
 #include "paddle/fluid/pir/dialect/operator/ir/manual_op.h"
 #include "paddle/fluid/pir/dialect/operator/trait/custom_vjp.h"
 #include "paddle/phi/core/infermeta_utils.h"
