+++ paddle/fluid/pir/dialect/op_generator/op_kerneltype_gen.py
@@ -44,7 +44,7 @@ OP_SUPPORT_TRANSFORM_CHECK_TEMPLATE = """
 
 OP_COMPLEX_PROMOTE_CHECK_TEMPLATE = """
   // deal complex_promote
-  if (framework::IsComplexType(expected_kernel_dtype)) {{
+  if (phi::IsComplexType(expected_kernel_dtype)) {{
     // only promote inputs’s types when contains complex input
     return tensor_dtype;
   }}
