+++ paddle/fluid/pir/dialect/operator/ir/manual_op.h
@@ -15,7 +15,6 @@
 #pragma once
 #include <vector>
 
-#include "paddle/fluid/framework/infershape_utils.h"
 #include "paddle/fluid/pir/dialect/operator/interface/decomp.h"
 #include "paddle/fluid/pir/dialect/operator/interface/get_kernel_type_for_var.h"
 #include "paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/infer_symbolic_shape.h"
