+++ paddle/fluid/pir/transforms/gpu/conv2d_add_fuse_pass.cc
@@ -16,6 +16,9 @@
 #include "paddle/fluid/pir/dialect/operator/ir/op_attribute.h"
 #include "paddle/fluid/pir/dialect/operator/ir/pd_op.h"
 #include "paddle/fluid/pir/drr/include/drr_pattern_base.h"
+#ifdef PADDLE_WITH_CUDA
+#include "paddle/phi/core/platform/device/gpu/gpu_info.h"
+#endif
 
 #include "paddle/fluid/pir/utils/general_functions.h"
 #include "paddle/pir/include/pass/pass.h"
